Refactors Inbox, deprecates JRuby, other fixes
- removes `Inbox` in favor of `Nylas`
- Simplifies gemspec and fixes #84
- Removes JRuby support
