+++ .gitignore
@@ -21,4 +21,3 @@ tags
 TAGS
 *.swp
 vendor/
-*.gemspec
