+++ lib/stream_handlers/event_machine.rb
