+++ lib/stream_handlers/simple_stream.rb
