+++ lib/version.rb
@@ -1,3 +1,3 @@
-module Inbox
+module Nylas
   VERSION = "3.0.0"
 end
