+++ spec/opensource_spec.rb
@@ -1,8 +1,8 @@
 require 'event'
 
-describe Inbox::APIAccount do
+describe Nylas::APIAccount do
   before (:each) do
-    @inbox = Inbox::API.new(nil, nil, nil, 'http://localhost:5555')
+    @inbox = Nylas::API.new(nil, nil, nil, 'http://localhost:5555')
 
     @base_url = "http://localhost:5555/accounts?limit=100&offset=0"
     stub_request(:get, @base_url).to_return(
