+++ spec/restful_model_spec.rb
@@ -1,12 +1,12 @@
 
 describe 'RestfulModel' do
   before (:each) do
-    @api = Inbox::API.new('app_id', 'app_secret', 'key')
+    @api = Nylas::API.new('app_id', 'app_secret', 'key')
   end
 
   describe "#as_json" do
     it "should return a hash with the parameter properties" do
-      r = Inbox::RestfulModel.new(@api)
+      r = Nylas::RestfulModel.new(@api)
       r.id = '1'
       r.account_id = '123';
       r.created_at = Time.new
@@ -14,7 +14,7 @@ describe 'RestfulModel' do
     end
 
     it "should ignore arbitrary setters" do
-      model_subclass = Class.new(Inbox::RestfulModel) do
+      model_subclass = Class.new(Nylas::RestfulModel) do
         def foo=(bar); end
       end
       r = model_subclass.new(@api)
@@ -28,7 +28,7 @@ describe 'RestfulModel' do
   describe "#inflate" do
     it "should set the values provided in the hash on the instance" do
       now = Time.new
-      r = Inbox::RestfulModel.new(@api)
+      r = Nylas::RestfulModel.new(@api)
       r.inflate({"id" => "1", "account_id" => "123", "created_at" => now})
 
       expect(r.id).to eq('1')
@@ -37,7 +37,7 @@ describe 'RestfulModel' do
     end
 
     it "should ignore arbitrary json values" do
-      model_subclass = Class.new(Inbox::RestfulModel) do
+      model_subclass = Class.new(Nylas::RestfulModel) do
         attr_accessor :foo
       end
 
@@ -49,7 +49,7 @@ describe 'RestfulModel' do
     it "should issue a DELETE when calling delete" do
       url = 'http://localhost:5555/messages/1'
       message_url = stub_request(:delete, url)
-      r = Inbox::RestfulModel.new(@api)
+      r = Nylas::RestfulModel.new(@api)
       allow(r).to receive_messages(:url => url)
 
       r.destroy
@@ -60,7 +60,7 @@ describe 'RestfulModel' do
       url = 'http://localhost:5555/events/1'
       stubbed_url = 'http://localhost:5555/events/1?param1=&param2=stuff&send_notifications=true'
       message_url = stub_request(:delete, stubbed_url)
-      r = Inbox::RestfulModel.new(@api)
+      r = Nylas::RestfulModel.new(@api)
       allow(r).to receive_messages(:url => url)
 
       r.destroy(:send_notifications => true, :param1 => nil, :param2 => 'stuff')
