+++ spec/thread_spec.rb
@@ -1,22 +1,22 @@
 require 'thread'
 require 'folder'
 
-describe Inbox::Thread do
+describe Nylas::Thread do
   before (:each) do
     @app_id = 'ABC'
     @app_secret = '123'
     @access_token = 'UXXMOCJW-BKSLPCFI-UQAQFWLO'
-    @inbox = Inbox::API.new(@app_id, @app_secret, @access_token)
+    @inbox = Nylas::API.new(@app_id, @app_secret, @access_token)
   end
 
   describe "#as_json" do
     it "only includes labels/folder info" do
-      thr = Inbox::Thread.new(@inbox, nil)
+      thr = Nylas::Thread.new(@inbox, nil)
       thr.subject = 'Test event'
 
       labels = ['test label', 'label 2']
       labels.map! do |label|
-        l = Inbox::Label.new(@inbox, nil)
+        l = Nylas::Label.new(@inbox, nil)
         l.id = label
         l
       end
@@ -27,7 +27,7 @@ describe Inbox::Thread do
       expect(dict['label_ids']).to eq(['test label', 'label 2'])
 
       # Now check that we do the same if @folder is set.
-      thr = Inbox::Thread.new(@inbox, nil)
+      thr = Nylas::Thread.new(@inbox, nil)
       thr.subject = 'Test event'
       thr.folder = labels[0]
       dict = thr.as_json
@@ -45,7 +45,7 @@ describe Inbox::Thread do
         with(basic_auth: [@access_token]).
         to_return(:status => 200, :body => '{"unread": false}')
 
-      th = Inbox::Thread.new(@inbox, nil)
+      th = Nylas::Thread.new(@inbox, nil)
       th.id = 2
       th.mark_as_read!
       expect(a_request(:put, url)).to have_been_made.once
@@ -60,7 +60,7 @@ describe Inbox::Thread do
         with(basic_auth: [@access_token]).
         to_return(:status => 200, :body => '{"starred": true}')
 
-      th = Inbox::Thread.new(@inbox, nil)
+      th = Nylas::Thread.new(@inbox, nil)
       th.id = 2
       th.star!
       expect(a_request(:put, url)).to have_been_made.once
